requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]

# SQLAlchemy ORM with Async Support
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0

# Database Migrations
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Configuration
pydantic-settings
email-validator

# Development
python-dotenv
httpx